refactor: changing currency to USD
