+++ lib/src/presenter/provider/visible_provider.dart
