+++ lib/utils/currency_formater.dart
@@ -1,3 +1,6 @@
 import 'package:intl/intl.dart';
 
-NumberFormat formatter = NumberFormat("#,##0.00", "pt-BR");
+NumberFormat formatter = NumberFormat.currency(
+  locale: "en_US",
+  symbol: "US\$ "
+);
